Collision start
Made some progress on collision but ended up making a bit of a mess. When walking against walls in a positive direction the collision is fine but it works worse against negative walls. I suspect this is because the collision is essentially checking two walls instead of the whole cube.